+++ TidalPy/tides/modes/mode_calc_helper/eccen_calc_orderl3.py
@@ -1,16 +1,16 @@
 from typing import Dict
 
-from ..eccentricity_funcs import EccenOutput, orderl2, orderl3, orderl4, orderl5
-from ...utilities.performance.numba import njit
-from ...utilities.types import FloatArray
+from ...eccentricity_funcs import EccenOutput, orderl2, orderl3
+from ....utilities.performance.numba import njit
+from ....utilities.types import FloatArray
 
 
 @njit(cacheable=True)
-def eccentricity_truncation_2_maxl_5(eccentricity: FloatArray) -> Dict[int, EccenOutput]:
+def eccentricity_truncation_2_maxl_3(eccentricity: FloatArray) -> Dict[int, EccenOutput]:
     """ Calculates eccentricity functions (squared) for a given maximum tidal order (going through each l)
 
     Truncation level = 2
-    Max Supported l = 5
+    Max Supported l = 3
 
     Parameters
     ----------
@@ -24,20 +24,18 @@ def eccentricity_truncation_2_maxl_5(eccentricity: FloatArray) -> Dict[int, Ecce
 
     result_by_orderl = {
         2: orderl2.eccentricity_funcs_trunc2(eccentricity),
-        3: orderl3.eccentricity_funcs_trunc2(eccentricity),
-        4: orderl4.eccentricity_funcs_trunc2(eccentricity),
-        5: orderl5.eccentricity_funcs_trunc2(eccentricity)
+        3: orderl3.eccentricity_funcs_trunc2(eccentricity)
         }
 
     return result_by_orderl
 
 
 @njit(cacheable=True)
-def eccentricity_truncation_4_maxl_5(eccentricity: FloatArray) -> Dict[int, EccenOutput]:
+def eccentricity_truncation_4_maxl_3(eccentricity: FloatArray) -> Dict[int, EccenOutput]:
     """ Calculates eccentricity functions (squared) for a given maximum tidal order (going through each l)
 
     Truncation level = 4
-    Max Supported l = 5
+    Max Supported l = 3
 
     Parameters
     ----------
@@ -51,20 +49,18 @@ def eccentricity_truncation_4_maxl_5(eccentricity: FloatArray) -> Dict[int, Ecce
 
     result_by_orderl = {
         2: orderl2.eccentricity_funcs_trunc4(eccentricity),
-        3: orderl3.eccentricity_funcs_trunc4(eccentricity),
-        4: orderl4.eccentricity_funcs_trunc4(eccentricity),
-        5: orderl5.eccentricity_funcs_trunc4(eccentricity)
+        3: orderl3.eccentricity_funcs_trunc4(eccentricity)
         }
 
     return result_by_orderl
 
 
 @njit(cacheable=True)
-def eccentricity_truncation_6_maxl_5(eccentricity: FloatArray) -> Dict[int, EccenOutput]:
+def eccentricity_truncation_6_maxl_3(eccentricity: FloatArray) -> Dict[int, EccenOutput]:
     """ Calculates eccentricity functions (squared) for a given maximum tidal order (going through each l)
 
     Truncation level = 6
-    Max Supported l = 5
+    Max Supported l = 3
 
     Parameters
     ----------
@@ -78,20 +74,18 @@ def eccentricity_truncation_6_maxl_5(eccentricity: FloatArray) -> Dict[int, Ecce
 
     result_by_orderl = {
         2: orderl2.eccentricity_funcs_trunc6(eccentricity),
-        3: orderl3.eccentricity_funcs_trunc6(eccentricity),
-        4: orderl4.eccentricity_funcs_trunc6(eccentricity),
-        5: orderl5.eccentricity_funcs_trunc6(eccentricity)
+        3: orderl3.eccentricity_funcs_trunc6(eccentricity)
         }
 
     return result_by_orderl
 
 
 @njit(cacheable=True)
-def eccentricity_truncation_8_maxl_5(eccentricity: FloatArray) -> Dict[int, EccenOutput]:
+def eccentricity_truncation_8_maxl_3(eccentricity: FloatArray) -> Dict[int, EccenOutput]:
     """ Calculates eccentricity functions (squared) for a given maximum tidal order (going through each l)
 
     Truncation level = 8
-    Max Supported l = 5
+    Max Supported l = 3
 
     Parameters
     ----------
@@ -105,20 +99,18 @@ def eccentricity_truncation_8_maxl_5(eccentricity: FloatArray) -> Dict[int, Ecce
 
     result_by_orderl = {
         2: orderl2.eccentricity_funcs_trunc8(eccentricity),
-        3: orderl3.eccentricity_funcs_trunc8(eccentricity),
-        4: orderl4.eccentricity_funcs_trunc8(eccentricity),
-        5: orderl5.eccentricity_funcs_trunc8(eccentricity)
+        3: orderl3.eccentricity_funcs_trunc8(eccentricity)
         }
 
     return result_by_orderl
 
 
 @njit(cacheable=True)
-def eccentricity_truncation_10_maxl_5(eccentricity: FloatArray) -> Dict[int, EccenOutput]:
+def eccentricity_truncation_10_maxl_3(eccentricity: FloatArray) -> Dict[int, EccenOutput]:
     """ Calculates eccentricity functions (squared) for a given maximum tidal order (going through each l)
 
     Truncation level = 10
-    Max Supported l = 5
+    Max Supported l = 3
 
     Parameters
     ----------
@@ -132,20 +124,18 @@ def eccentricity_truncation_10_maxl_5(eccentricity: FloatArray) -> Dict[int, Ecc
 
     result_by_orderl = {
         2: orderl2.eccentricity_funcs_trunc10(eccentricity),
-        3: orderl3.eccentricity_funcs_trunc10(eccentricity),
-        4: orderl4.eccentricity_funcs_trunc10(eccentricity),
-        5: orderl5.eccentricity_funcs_trunc10(eccentricity)
+        3: orderl3.eccentricity_funcs_trunc10(eccentricity)
         }
 
     return result_by_orderl
 
 
 @njit(cacheable=True)
-def eccentricity_truncation_12_maxl_5(eccentricity: FloatArray) -> Dict[int, EccenOutput]:
+def eccentricity_truncation_12_maxl_3(eccentricity: FloatArray) -> Dict[int, EccenOutput]:
     """ Calculates eccentricity functions (squared) for a given maximum tidal order (going through each l)
 
     Truncation level = 12
-    Max Supported l = 5
+    Max Supported l = 3
 
     Parameters
     ----------
@@ -159,20 +149,18 @@ def eccentricity_truncation_12_maxl_5(eccentricity: FloatArray) -> Dict[int, Ecc
 
     result_by_orderl = {
         2: orderl2.eccentricity_funcs_trunc12(eccentricity),
-        3: orderl3.eccentricity_funcs_trunc12(eccentricity),
-        4: orderl4.eccentricity_funcs_trunc12(eccentricity),
-        5: orderl5.eccentricity_funcs_trunc12(eccentricity)
+        3: orderl3.eccentricity_funcs_trunc12(eccentricity)
         }
 
     return result_by_orderl
 
 
 @njit(cacheable=True)
-def eccentricity_truncation_14_maxl_5(eccentricity: FloatArray) -> Dict[int, EccenOutput]:
+def eccentricity_truncation_14_maxl_3(eccentricity: FloatArray) -> Dict[int, EccenOutput]:
     """ Calculates eccentricity functions (squared) for a given maximum tidal order (going through each l)
 
     Truncation level = 14
-    Max Supported l = 5
+    Max Supported l = 3
 
     Parameters
     ----------
@@ -186,20 +174,18 @@ def eccentricity_truncation_14_maxl_5(eccentricity: FloatArray) -> Dict[int, Ecc
 
     result_by_orderl = {
         2: orderl2.eccentricity_funcs_trunc14(eccentricity),
-        3: orderl3.eccentricity_funcs_trunc14(eccentricity),
-        4: orderl4.eccentricity_funcs_trunc14(eccentricity),
-        5: orderl5.eccentricity_funcs_trunc14(eccentricity)
+        3: orderl3.eccentricity_funcs_trunc14(eccentricity)
         }
 
     return result_by_orderl
 
 
 @njit(cacheable=True)
-def eccentricity_truncation_16_maxl_5(eccentricity: FloatArray) -> Dict[int, EccenOutput]:
+def eccentricity_truncation_16_maxl_3(eccentricity: FloatArray) -> Dict[int, EccenOutput]:
     """ Calculates eccentricity functions (squared) for a given maximum tidal order (going through each l)
 
     Truncation level = 16
-    Max Supported l = 5
+    Max Supported l = 3
 
     Parameters
     ----------
@@ -213,20 +199,18 @@ def eccentricity_truncation_16_maxl_5(eccentricity: FloatArray) -> Dict[int, Ecc
 
     result_by_orderl = {
         2: orderl2.eccentricity_funcs_trunc16(eccentricity),
-        3: orderl3.eccentricity_funcs_trunc16(eccentricity),
-        4: orderl4.eccentricity_funcs_trunc16(eccentricity),
-        5: orderl5.eccentricity_funcs_trunc16(eccentricity)
+        3: orderl3.eccentricity_funcs_trunc16(eccentricity)
         }
 
     return result_by_orderl
 
 
 @njit(cacheable=True)
-def eccentricity_truncation_18_maxl_5(eccentricity: FloatArray) -> Dict[int, EccenOutput]:
+def eccentricity_truncation_18_maxl_3(eccentricity: FloatArray) -> Dict[int, EccenOutput]:
     """ Calculates eccentricity functions (squared) for a given maximum tidal order (going through each l)
 
     Truncation level = 18
-    Max Supported l = 5
+    Max Supported l = 3
 
     Parameters
     ----------
@@ -240,20 +224,18 @@ def eccentricity_truncation_18_maxl_5(eccentricity: FloatArray) -> Dict[int, Ecc
 
     result_by_orderl = {
         2: orderl2.eccentricity_funcs_trunc18(eccentricity),
-        3: orderl3.eccentricity_funcs_trunc18(eccentricity),
-        4: orderl4.eccentricity_funcs_trunc18(eccentricity),
-        5: orderl5.eccentricity_funcs_trunc18(eccentricity)
+        3: orderl3.eccentricity_funcs_trunc18(eccentricity)
         }
 
     return result_by_orderl
 
 
 @njit(cacheable=True)
-def eccentricity_truncation_20_maxl_5(eccentricity: FloatArray) -> Dict[int, EccenOutput]:
+def eccentricity_truncation_20_maxl_3(eccentricity: FloatArray) -> Dict[int, EccenOutput]:
     """ Calculates eccentricity functions (squared) for a given maximum tidal order (going through each l)
 
     Truncation level = 20
-    Max Supported l = 5
+    Max Supported l = 3
 
     Parameters
     ----------
@@ -267,9 +249,7 @@ def eccentricity_truncation_20_maxl_5(eccentricity: FloatArray) -> Dict[int, Ecc
 
     result_by_orderl = {
         2: orderl2.eccentricity_funcs_trunc20(eccentricity),
-        3: orderl3.eccentricity_funcs_trunc20(eccentricity),
-        4: orderl4.eccentricity_funcs_trunc20(eccentricity),
-        5: orderl5.eccentricity_funcs_trunc20(eccentricity)
+        3: orderl3.eccentricity_funcs_trunc20(eccentricity)
         }
 
     return result_by_orderl