+++ TidalPy/tides/modes/mode_calc_helper/eccen_calc_orderl6.py
@@ -1,16 +1,16 @@
 from typing import Dict
 
-from ..eccentricity_funcs import EccenOutput, orderl2, orderl3, orderl4
-from ...utilities.performance.numba import njit
-from ...utilities.types import FloatArray
+from ...eccentricity_funcs import EccenOutput, orderl2, orderl3, orderl4, orderl5, orderl6
+from ....utilities.performance.numba import njit
+from ....utilities.types import FloatArray
 
 
 @njit(cacheable=True)
-def eccentricity_truncation_2_maxl_4(eccentricity: FloatArray) -> Dict[int, EccenOutput]:
+def eccentricity_truncation_2_maxl_6(eccentricity: FloatArray) -> Dict[int, EccenOutput]:
     """ Calculates eccentricity functions (squared) for a given maximum tidal order (going through each l)
 
     Truncation level = 2
-    Max Supported l = 4
+    Max Supported l = 6
 
     Parameters
     ----------
@@ -25,18 +25,20 @@ def eccentricity_truncation_2_maxl_4(eccentricity: FloatArray) -> Dict[int, Ecce
     result_by_orderl = {
         2: orderl2.eccentricity_funcs_trunc2(eccentricity),
         3: orderl3.eccentricity_funcs_trunc2(eccentricity),
-        4: orderl4.eccentricity_funcs_trunc2(eccentricity)
+        4: orderl4.eccentricity_funcs_trunc2(eccentricity),
+        5: orderl5.eccentricity_funcs_trunc2(eccentricity),
+        6: orderl6.eccentricity_funcs_trunc2(eccentricity)
         }
 
     return result_by_orderl
 
 
 @njit(cacheable=True)
-def eccentricity_truncation_4_maxl_4(eccentricity: FloatArray) -> Dict[int, EccenOutput]:
+def eccentricity_truncation_4_maxl_6(eccentricity: FloatArray) -> Dict[int, EccenOutput]:
     """ Calculates eccentricity functions (squared) for a given maximum tidal order (going through each l)
 
     Truncation level = 4
-    Max Supported l = 4
+    Max Supported l = 6
 
     Parameters
     ----------
@@ -51,18 +53,20 @@ def eccentricity_truncation_4_maxl_4(eccentricity: FloatArray) -> Dict[int, Ecce
     result_by_orderl = {
         2: orderl2.eccentricity_funcs_trunc4(eccentricity),
         3: orderl3.eccentricity_funcs_trunc4(eccentricity),
-        4: orderl4.eccentricity_funcs_trunc4(eccentricity)
+        4: orderl4.eccentricity_funcs_trunc4(eccentricity),
+        5: orderl5.eccentricity_funcs_trunc4(eccentricity),
+        6: orderl6.eccentricity_funcs_trunc4(eccentricity)
         }
 
     return result_by_orderl
 
 
 @njit(cacheable=True)
-def eccentricity_truncation_6_maxl_4(eccentricity: FloatArray) -> Dict[int, EccenOutput]:
+def eccentricity_truncation_6_maxl_6(eccentricity: FloatArray) -> Dict[int, EccenOutput]:
     """ Calculates eccentricity functions (squared) for a given maximum tidal order (going through each l)
 
     Truncation level = 6
-    Max Supported l = 4
+    Max Supported l = 6
 
     Parameters
     ----------
@@ -77,18 +81,20 @@ def eccentricity_truncation_6_maxl_4(eccentricity: FloatArray) -> Dict[int, Ecce
     result_by_orderl = {
         2: orderl2.eccentricity_funcs_trunc6(eccentricity),
         3: orderl3.eccentricity_funcs_trunc6(eccentricity),
-        4: orderl4.eccentricity_funcs_trunc6(eccentricity)
+        4: orderl4.eccentricity_funcs_trunc6(eccentricity),
+        5: orderl5.eccentricity_funcs_trunc6(eccentricity),
+        6: orderl6.eccentricity_funcs_trunc6(eccentricity)
         }
 
     return result_by_orderl
 
 
 @njit(cacheable=True)
-def eccentricity_truncation_8_maxl_4(eccentricity: FloatArray) -> Dict[int, EccenOutput]:
+def eccentricity_truncation_8_maxl_6(eccentricity: FloatArray) -> Dict[int, EccenOutput]:
     """ Calculates eccentricity functions (squared) for a given maximum tidal order (going through each l)
 
     Truncation level = 8
-    Max Supported l = 4
+    Max Supported l = 6
 
     Parameters
     ----------
@@ -103,18 +109,20 @@ def eccentricity_truncation_8_maxl_4(eccentricity: FloatArray) -> Dict[int, Ecce
     result_by_orderl = {
         2: orderl2.eccentricity_funcs_trunc8(eccentricity),
         3: orderl3.eccentricity_funcs_trunc8(eccentricity),
-        4: orderl4.eccentricity_funcs_trunc8(eccentricity)
+        4: orderl4.eccentricity_funcs_trunc8(eccentricity),
+        5: orderl5.eccentricity_funcs_trunc8(eccentricity),
+        6: orderl6.eccentricity_funcs_trunc8(eccentricity)
         }
 
     return result_by_orderl
 
 
 @njit(cacheable=True)
-def eccentricity_truncation_10_maxl_4(eccentricity: FloatArray) -> Dict[int, EccenOutput]:
+def eccentricity_truncation_10_maxl_6(eccentricity: FloatArray) -> Dict[int, EccenOutput]:
     """ Calculates eccentricity functions (squared) for a given maximum tidal order (going through each l)
 
     Truncation level = 10
-    Max Supported l = 4
+    Max Supported l = 6
 
     Parameters
     ----------
@@ -129,18 +137,20 @@ def eccentricity_truncation_10_maxl_4(eccentricity: FloatArray) -> Dict[int, Ecc
     result_by_orderl = {
         2: orderl2.eccentricity_funcs_trunc10(eccentricity),
         3: orderl3.eccentricity_funcs_trunc10(eccentricity),
-        4: orderl4.eccentricity_funcs_trunc10(eccentricity)
+        4: orderl4.eccentricity_funcs_trunc10(eccentricity),
+        5: orderl5.eccentricity_funcs_trunc10(eccentricity),
+        6: orderl6.eccentricity_funcs_trunc10(eccentricity)
         }
 
     return result_by_orderl
 
 
 @njit(cacheable=True)
-def eccentricity_truncation_12_maxl_4(eccentricity: FloatArray) -> Dict[int, EccenOutput]:
+def eccentricity_truncation_12_maxl_6(eccentricity: FloatArray) -> Dict[int, EccenOutput]:
     """ Calculates eccentricity functions (squared) for a given maximum tidal order (going through each l)
 
     Truncation level = 12
-    Max Supported l = 4
+    Max Supported l = 6
 
     Parameters
     ----------
@@ -155,18 +165,20 @@ def eccentricity_truncation_12_maxl_4(eccentricity: FloatArray) -> Dict[int, Ecc
     result_by_orderl = {
         2: orderl2.eccentricity_funcs_trunc12(eccentricity),
         3: orderl3.eccentricity_funcs_trunc12(eccentricity),
-        4: orderl4.eccentricity_funcs_trunc12(eccentricity)
+        4: orderl4.eccentricity_funcs_trunc12(eccentricity),
+        5: orderl5.eccentricity_funcs_trunc12(eccentricity),
+        6: orderl6.eccentricity_funcs_trunc12(eccentricity)
         }
 
     return result_by_orderl
 
 
 @njit(cacheable=True)
-def eccentricity_truncation_14_maxl_4(eccentricity: FloatArray) -> Dict[int, EccenOutput]:
+def eccentricity_truncation_14_maxl_6(eccentricity: FloatArray) -> Dict[int, EccenOutput]:
     """ Calculates eccentricity functions (squared) for a given maximum tidal order (going through each l)
 
     Truncation level = 14
-    Max Supported l = 4
+    Max Supported l = 6
 
     Parameters
     ----------
@@ -181,18 +193,20 @@ def eccentricity_truncation_14_maxl_4(eccentricity: FloatArray) -> Dict[int, Ecc
     result_by_orderl = {
         2: orderl2.eccentricity_funcs_trunc14(eccentricity),
         3: orderl3.eccentricity_funcs_trunc14(eccentricity),
-        4: orderl4.eccentricity_funcs_trunc14(eccentricity)
+        4: orderl4.eccentricity_funcs_trunc14(eccentricity),
+        5: orderl5.eccentricity_funcs_trunc14(eccentricity),
+        6: orderl6.eccentricity_funcs_trunc14(eccentricity)
         }
 
     return result_by_orderl
 
 
 @njit(cacheable=True)
-def eccentricity_truncation_16_maxl_4(eccentricity: FloatArray) -> Dict[int, EccenOutput]:
+def eccentricity_truncation_16_maxl_6(eccentricity: FloatArray) -> Dict[int, EccenOutput]:
     """ Calculates eccentricity functions (squared) for a given maximum tidal order (going through each l)
 
     Truncation level = 16
-    Max Supported l = 4
+    Max Supported l = 6
 
     Parameters
     ----------
@@ -207,18 +221,20 @@ def eccentricity_truncation_16_maxl_4(eccentricity: FloatArray) -> Dict[int, Ecc
     result_by_orderl = {
         2: orderl2.eccentricity_funcs_trunc16(eccentricity),
         3: orderl3.eccentricity_funcs_trunc16(eccentricity),
-        4: orderl4.eccentricity_funcs_trunc16(eccentricity)
+        4: orderl4.eccentricity_funcs_trunc16(eccentricity),
+        5: orderl5.eccentricity_funcs_trunc16(eccentricity),
+        6: orderl6.eccentricity_funcs_trunc16(eccentricity)
         }
 
     return result_by_orderl
 
 
 @njit(cacheable=True)
-def eccentricity_truncation_18_maxl_4(eccentricity: FloatArray) -> Dict[int, EccenOutput]:
+def eccentricity_truncation_18_maxl_6(eccentricity: FloatArray) -> Dict[int, EccenOutput]:
     """ Calculates eccentricity functions (squared) for a given maximum tidal order (going through each l)
 
     Truncation level = 18
-    Max Supported l = 4
+    Max Supported l = 6
 
     Parameters
     ----------
@@ -233,18 +249,20 @@ def eccentricity_truncation_18_maxl_4(eccentricity: FloatArray) -> Dict[int, Ecc
     result_by_orderl = {
         2: orderl2.eccentricity_funcs_trunc18(eccentricity),
         3: orderl3.eccentricity_funcs_trunc18(eccentricity),
-        4: orderl4.eccentricity_funcs_trunc18(eccentricity)
+        4: orderl4.eccentricity_funcs_trunc18(eccentricity),
+        5: orderl5.eccentricity_funcs_trunc18(eccentricity),
+        6: orderl6.eccentricity_funcs_trunc18(eccentricity)
         }
 
     return result_by_orderl
 
 
 @njit(cacheable=True)
-def eccentricity_truncation_20_maxl_4(eccentricity: FloatArray) -> Dict[int, EccenOutput]:
+def eccentricity_truncation_20_maxl_6(eccentricity: FloatArray) -> Dict[int, EccenOutput]:
     """ Calculates eccentricity functions (squared) for a given maximum tidal order (going through each l)
 
     Truncation level = 20
-    Max Supported l = 4
+    Max Supported l = 6
 
     Parameters
     ----------
@@ -259,7 +277,9 @@ def eccentricity_truncation_20_maxl_4(eccentricity: FloatArray) -> Dict[int, Ecc
     result_by_orderl = {
         2: orderl2.eccentricity_funcs_trunc20(eccentricity),
         3: orderl3.eccentricity_funcs_trunc20(eccentricity),
-        4: orderl4.eccentricity_funcs_trunc20(eccentricity)
+        4: orderl4.eccentricity_funcs_trunc20(eccentricity),
+        5: orderl5.eccentricity_funcs_trunc20(eccentricity),
+        6: orderl6.eccentricity_funcs_trunc20(eccentricity)
         }
 
     return result_by_orderl