+++ TidalPy/tides/modes/mode_calc_helper/inclin_calc_orderl3.py
@@ -1,17 +1,17 @@
 from typing import Dict
 
-from ..inclination_funcs import InclinOutput, orderl2, orderl3, orderl4, orderl5
-from ...utilities.performance.numba import njit
-from ...utilities.types import FloatArray
+from ...inclination_funcs import InclinOutput, orderl2, orderl3
+from ....utilities.performance.numba import njit
+from ....utilities.types import FloatArray
 
 
 @njit(cacheable=True)
-def inclination_off_maxl_5(obliquity: FloatArray) -> Dict[int, InclinOutput]:
+def inclination_off_maxl_3(obliquity: FloatArray) -> Dict[int, InclinOutput]:
     """ Calculates inclination functions (squared) for a given maximum tidal order (going through each l) - Off Mode
 
     Obliquity is assumed to be zero.
 
-    Max Supported l = 5
+    Max Supported l = 3
 
     Parameters
     ----------
@@ -25,21 +25,19 @@ def inclination_off_maxl_5(obliquity: FloatArray) -> Dict[int, InclinOutput]:
 
     result_by_orderl = {
         2: orderl2.calc_inclination_off(obliquity),
-        3: orderl3.calc_inclination_off(obliquity),
-        4: orderl4.calc_inclination_off(obliquity),
-        5: orderl5.calc_inclination_off(obliquity)
+        3: orderl3.calc_inclination_off(obliquity)
         }
 
     return result_by_orderl
 
 
 @njit(cacheable=True)
-def inclination_on_maxl_5(obliquity: FloatArray) -> Dict[int, InclinOutput]:
+def inclination_on_maxl_3(obliquity: FloatArray) -> Dict[int, InclinOutput]:
     """ Calculates inclination functions (squared) for a given maximum tidal order (going through each l) - On Mode
 
     Obliquity can be arbitrary.
 
-    Max Supported l = 5
+    Max Supported l = 3
 
     Parameters
     ----------
@@ -53,9 +51,7 @@ def inclination_on_maxl_5(obliquity: FloatArray) -> Dict[int, InclinOutput]:
 
     result_by_orderl = {
         2: orderl2.calc_inclination(obliquity),
-        3: orderl3.calc_inclination(obliquity),
-        4: orderl4.calc_inclination(obliquity),
-        5: orderl5.calc_inclination(obliquity)
+        3: orderl3.calc_inclination(obliquity)
         }
 
     return result_by_orderl